This commit handels the	error from push script

+++ build_image.sh
@@ -98,6 +98,10 @@ cd docker && \
  sudo docker build -f ${DOCKERFILE} -t ${REPO_NAME}:ci ${DBUILD_ARGS} . && \
  DIMAGE=${REPO_NAME} ./push && \
  cd ..
+if [ $? -ne 0 ]; then
+ echo "Failed to run push script"
+ exit 1
+fi
 
 rm -rf ./docker/istgt*
